refactor: split out orderedMerge and PipelineState
